Moved a step that referenced the prometheus installation instructions to after our simple installation.
Previously we had this broken into a simple series of steps to install prometheus and a series of steps to configure prometheus. When these two sets of steps were combined into one, they had a step in the middle to install prometheus. However, installing prometheus was already done by the prior steps and did not need its own section. The reference to Prometheus's documentation was included to point out alternate ways of installing to more savvy users.