Corrections to wiring guide (#13)
* Corrections to wiring guide

The SM_1014S system at the moment is a 1 NIC system. Hence we have unified networking. There is no concept today at Qumulo of isolating traffic to individual ports on a NIC. We only isolate traffic between NICs. So for a 1 NIC system, all traffic, client and node goes through the same NIC and the ports on that NIC. Hence a SM_1014S system could be said to have a Unified FrontEnd or BackEnd NIC. In the code this is expressed as the system only has a FrontEnd NIC, and no Backend NIC, hence all traffic goes through the 1 NIC.

* Add back in the split wiring recommendation. 

To maximize redundancy, split interfaces across subnets by connecting each port to a different switch.

* Minor correction and insertion of var.

* Simplifying for var.

* Removing minor correction.

Co-authored-by: Michael G. Khmelnitsky <mkhmelnitsky@qumulo.com>